Moved the API routes to routes folder. Corrected syntax and made minor changes for readability and conciseness.
